+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## Version 0.56.3
 
-- Fixed UNREGISTERED_CHAIN_MAX_LEN const for Venom blockchain.
+- Supported new multi-networks API of DHT
+- United crate for protocols
 
 ## Version 0.56.2